Function post_expansion::strip_attrs_now

pub fn strip_attrs_now(ast: MacroInput, strip: &[&str]) -> MacroInput

Strip the specified attributes from the AST. Note that if it is possible those attributes may be needed by other custom derives, you should use strip_attrs_later instead which strips the attributes only after other custom derives have had a chance to see them.

ast = post_expansion::strip_attrs_now(ast, &["serde"]);