Macro post_expansion::register_post_expansion

macro_rules! register_post_expansion {
    ($id:ident) => { ... };
}

Register a Macros 1.1 custom derive mode corresponding to a post-expansion pass for the current crate. Must be called at the top level of a proc-macro crate. Must be called with an identifier for the post-expansion pass, which must be "PostExpansion_" + unique identifier. Recommended to use the crate name of the current crate as the unique identifier.

Only needed if using strip_attrs_later. The identifier passed to strip_attrs_later must match the unique identifier here.

register_post_expansion!(PostExpansion_elastic_types);